dangtm

+++ AppService/TayHo.ProjectManager/PrjectManager.CMD/ProjectManager.CMD.Api/Application/Commands/NS_CongViecDetailCmd/BaseClasses/NS_CongViecDetailCommandSet.cs
@@ -7,6 +7,6 @@ namespace ProjectManager.CMD.Api.Application.Commands
         public int? GiaiDoanId { get; set; }
         public int? ReasonId { get; set; }
         public decimal? DonGia { get; set; }
-        public int? KhoiLuong { get; set; }
+        public decimal? KhoiLuong { get; set; }
     }
 }
