+++ AppService/TayHo.ProjectManager/PrjectManager.Read/ProjectManager.Read.Api/ViewModels/NS_CongViecResponseViewModel.cs
@@ -18,6 +18,6 @@ namespace ProjectManager.Read.Api.ViewModels
         public int? GiaiDoanId { get; set; }
         public int? ReasonId { get; set; }
         public decimal? DonGia { get; set; }
-        public int? KhoiLuong { get; set; }
+        public decimal? KhoiLuong { get; set; }
     }
 }
